backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
langchain
langchain-openai
openai
PyJWT
aiohttp
PyMuPDF
python-multipart
pydantic